fix: OOB and hangs in version/printf parsing
- pkg_version: fix the "snap" stage entry, whose namelen was 6 for a
  4-char name, causing a heap OOB read and dangling pointer in
  get_component on versions ending in "snap". Set namelen to 4 and
  value to "s"-"a"+ABASE so snap sorts after alpha as intended (#2691)
- pkg_printf: stop format_code from reading f[1] past end of string,
  and emit a lone "%" literally instead of looping forever when no
  valid format code follows
- event: cast nbdigits to int for the %* field width
- fetch_file: print the offending URL in the "hostname too long"
  error message

+++ src/event.c
@@ -195,7 +195,7 @@ job_status_begin(xstring *msg)
 	if ((nbtodl > 0 || nbactions > 0) && nbdone > 0) {
 		if (nbdigits == 0)
 			nbdigits = count_digits(nbtodl ? nbtodl : nbactions);
-		xstring_printf(msg, "[%*zu/%zu] ", nbdigits, nbdone, (nbtodl) ? nbtodl : nbactions);
+		xstring_printf(msg, "[%*zu/%zu] ", (int)nbdigits, nbdone, (nbtodl) ? nbtodl : nbactions);
 	}
 	if (nbtodl > 0 && nbtodl == nbdone) {
 		nbtodl = 0;
